add all sorting methods
