--- server/include/select.h
@@ -1,3 +1,0 @@
-#include <vector>
-using namespace std;
-void selectionSort(vector<int>);
