+++ server/include/sorts.h
@@ -1,0 +1,8 @@
+#ifndef sorts_h
+#define sorts_h
+
+#include <vector>
+using namespace std;
+void selectionSort(vector<int>& arr);
+void shellSort(vector<int>& arr);
+#endif //sorts_h
